requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
chromadb
langchain
langgraph
discogs-client
python-dotenv
